+++ src/openpi/models_pytorch/transformers_replace/models/siglip/modeling_siglip.py
@@ -773,7 +773,7 @@ class SiglipVisionTransformer(nn.Module):
         )
 
         hidden_states = self.embeddings(pixel_values, interpolate_pos_encoding=interpolate_pos_encoding)
-        hidden_states = hidden_states.to(torch.bfloat16)
+        #hidden_states = hidden_states.to(torch.bfloat16)
 
         encoder_outputs: BaseModelOutput = self.encoder(
             inputs_embeds=hidden_states,
